In Firewall and StaticFlowEntryManager, some REST APIs are added and modified to handle more exceptions.
